Add Ed25519 signature verification step to detect and avoid outputting faulty signatures (#11)
EdDSA is known to be vulnerable to fault attacks which can lead to secret key extraction if
two signatures over the same data can be collected.
Randomly occurring bitflips in specific parts of the computation might in principle result
in vulnerable faulty signatures being generated, hence we add the option to verify the signatures
before outputting them.

This commit also adds the `checkEdDSAFaultySignatures` flag to the global config
to be able to control the deployment of the eddsa check.
Support is limited to the global config object as the affected functions
currently don't take a config input, and we don't need to selectively enable
the option anyway, so we limit the scope of the changes.