Implement OpenPGP message grammar validation (add `config.enforceGrammar`)
It enforces a message structure as defined in
https://www.rfc-editor.org/rfc/rfc9580.html#section-10.3
(but slightly more permissive with Padding packets allowed in all cases).
Since we are unclear on whether this change might
impact handling of some messages in the wild, generated by
odd use-cases or non-conformant implementations, we
also add the option to disable the grammar check via
`config.enforceGrammar`.

This solution gives us the option to probe whether the
grammar is too disruptive in a real-world setting,
by testing it in the Proton ecosystem.

GrammarErrors are only sensitive in the context of
unauthenticated decrypted streams.