TS: emit type declarations for .ts files with external exports, and publish them under `dist/types`
This change makes it possible to easily expose types from internal .ts files.
This is achieved by adding a TS compilation step which generates .d.ts files from
.ts ones.
Additionally, a copying step is needed for existing .d.ts files, which are not
automatically moved to the outDir by tsc.
The dist types will still only include those exported by the index files.

Other tried approaches (to e.g. avoid the manual copying step) that were not viable:
- generating a single .d.ts bundle is not supported by tsc, and other tools/plugins
  are either deprecated or (in the case of `rollup-plugin-dts`) failed to process our
  source code due to unsupported JS syntax
- relying on the `rollup-typescript` plugin to emit declarations felt overly messy
  to configure and potentially unreliable as it generated declarations for .js files as well
- renaming existing .d.ts to .ts (alongside .js ones) is not an option since it results
in an empty output for the corresponding modules on Rollup compilation, as the .js
source files are ignored.