Add `config.maxArgon2MemoryExponent` for argon2 memory limit  (https://github.com/openpgpjs/openpgpjs/pull/1943)
Add `config.maxArgon2MemoryExponent` for argon2 memory limit

This limit is applied both on encryption (if `config.s2kType` is set to `enums.s2k.argon2`)
and decryption.
If the input memory exponent exceeds this value, the library will not attempt the argon2 key
derivation and instead directly throw an `Argon2OutOfMemoryError` error.